requirements.txt
pandas==2.2.3
python-dotenv==1.0.1
SQLAlchemy==2.0.38
psycopg2==2.9.10
